Modified some data types.
I did not manage to link the library without an actual implementation of the derived classes in GmshReaderException. It has something to do with the macro in the virtual what() function.

Physical names and ids are not mandatory in gmsh. Do we want to keep it very general or make some assumptions of how the msh should look like?

Going to use the issue tracker next time :-)